+++ deploy/base/kustomization.yaml
@@ -33,7 +33,10 @@ commonLabels:
 
 resources:
   - crd.yaml
-  - rbac.yaml
+  - ns.yaml
+  - service_account.yaml
+  - role.yaml
+  - role_binding.yaml
   - operator.yaml
   - config.yaml
