--- deploy/base/rbac.yaml
@@ -1,172 +1,0 @@
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: security-profiles-operator
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: security-profiles-operator
-  namespace: security-profiles-operator
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: config-map-reader
-  namespace: security-profiles-operator
-rules:
-- apiGroups:
-  - ""
-  resources:
-  - "configmaps"
-  - "pods"
-  - "nodes"
-  verbs:
-  - "get"
-  - "watch"
-  - "list"
-- apiGroups:
-  - ""
-  resources:
-  - "events"
-  verbs:
-  - "create"
-  - "patch"
-- apiGroups:
-  - apps
-  resources:
-  - deployments
-  - daemonsets
-  - replicasets
-  - statefulsets
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - "security-profiles-operator.x-k8s.io"
-  resources:
-  - "seccompprofiles"
-  - "selinuxpolicies"
-  verbs:
-  - "get"
-  - "watch"
-  - "list"
-  - "update"
-  - "patch"
-- apiGroups:
-  - "security-profiles-operator.x-k8s.io"
-  resources:
-  - "seccompprofiles/status"
-  - "selinuxpolicies/status"
-  verbs: 
-  - "get"
-  - "update"
-  - "patch"
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: config-map-reader-binding
-  namespace: security-profiles-operator
-subjects:
-- kind: ServiceAccount
-  name: security-profiles-operator
-  namespace: security-profiles-operator
-roleRef:
-  kind: ClusterRole
-  name: config-map-reader
-  apiGroup: rbac.authorization.k8s.io
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: Role
-metadata:
-  labels:
-    app: security-profiles-operator
-  name: security-profiles-operator
-  namespace: security-profiles-operator
-rules:
-- apiGroups:
-  - ""
-  resources:
-  - pods
-  - services
-  - services/finalizers
-  - endpoints
-  - persistentvolumeclaims
-  - events
-  - configmaps
-  - configmaps/finalizers
-  - secrets
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - apps
-  resources:
-  - deployments
-  - daemonsets
-  - replicasets
-  - statefulsets
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - apps
-  resourceNames:
-  - selinux-operator
-  resources:
-  - deployments/finalizers
-  verbs:
-  - update
-- apiGroups:
-  - ""
-  resources:
-  - pods
-  verbs:
-  - get
-- apiGroups:
-  - apps
-  resources:
-  - replicasets
-  - deployments
-  verbs:
-  - get
-- apiGroups:                  # Needed for using privileged containers
-  - security.openshift.io
-  resources:
-  - securitycontextconstraints
-  resourceNames:
-  - privileged
-  verbs:
-  - use
----
-kind: RoleBinding
-apiVersion: rbac.authorization.k8s.io/v1
-metadata:
-  labels:
-    app: security-profiles-operator
-  name: selinux-operator
-  namespace: security-profiles-operator
-subjects:
-- kind: ServiceAccount
-  name: security-profiles-operator
-roleRef:
-  kind: Role
-  name: security-profiles-operator
-  apiGroup: rbac.authorization.k8s.io
