+++ deploy/base/role_binding.yaml
@@ -1,0 +1,58 @@
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: spo-manager
+  namespace: security-profiles-operator
+subjects:
+- kind: ServiceAccount
+  name: spo-manager
+  namespace: security-profiles-operator
+roleRef:
+  kind: ClusterRole
+  name: spo-manager
+  apiGroup: rbac.authorization.k8s.io
+---
+kind: RoleBinding
+apiVersion: rbac.authorization.k8s.io/v1
+metadata:
+  labels:
+    app: security-profiles-operator
+  name: spo-manager
+  namespace: security-profiles-operator
+subjects:
+- kind: ServiceAccount
+  name: spo-manager
+roleRef:
+  kind: Role
+  name: spo-manager
+  apiGroup: rbac.authorization.k8s.io
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: spod
+  namespace: security-profiles-operator
+subjects:
+- kind: ServiceAccount
+  name: spod
+  namespace: security-profiles-operator
+roleRef:
+  kind: ClusterRole
+  name: spod
+  apiGroup: rbac.authorization.k8s.io
+---
+kind: RoleBinding
+apiVersion: rbac.authorization.k8s.io/v1
+metadata:
+  labels:
+    app: security-profiles-operator
+  name: spod
+  namespace: security-profiles-operator
+subjects:
+- kind: ServiceAccount
+  name: spod
+roleRef:
+  kind: Role
+  name: spod
+  apiGroup: rbac.authorization.k8s.io
