+++ deploy/overlays/namespaced/spo-manager-clusterrole.yaml
@@ -2,5 +2,8 @@
   path: /kind
   value: Role
 - op: replace
+  path: /metadata/name
+  value: spo-object-manager
+- op: replace
   path: /metadata/namespace
   value: NS_REPLACE
