+++ deploy/overlays/namespaced/spod-clusterrolebinding.yaml
@@ -1,0 +1,12 @@
+- op: replace
+  path: /kind
+  value: RoleBinding
+- op: replace
+  path: /metadata/name
+  value: spod-object-manager
+- op: replace
+  path: /metadata/namespace
+  value: NS_REPLACE
+- op: replace
+  path: /roleRef/kind
+  value: Role
