+++ internal/pkg/manager/selinuxpolicy/setup.go
