+++ internal/pkg/manager/selinuxpolicy/utils.go
