+++ test/tc_default_profiles_test.go
@@ -24,7 +24,7 @@ import (
 
 	"sigs.k8s.io/security-profiles-operator/api/seccompprofile/v1alpha1"
 	"sigs.k8s.io/security-profiles-operator/internal/pkg/config"
-	"sigs.k8s.io/security-profiles-operator/internal/pkg/controllers/profile"
+	"sigs.k8s.io/security-profiles-operator/internal/pkg/daemon/profile"
 )
 
 func (e *e2e) testCaseDefaultAndExampleProfiles(nodes []string) {
